requirements.txt
python-bitcoinrpc
flask
waitress